socialapp/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
filelock==3.0.12
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.3
virtualenv==20.4.2
wrapt==1.12.1
zope.interface==5.3.0
